Reset FirebaseMessagingService
