+++ src/main/java/com/server/EZY/notification/service/FirebaseMessagingService.java
@@ -1,23 +1,10 @@
 package com.server.EZY.notification.service;
 
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import com.google.firebase.messaging.FirebaseMessaging;
-import org.springframework.context.annotation.Bean;
-import org.springframework.core.io.ClassPathResource;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-import java.io.IOException;
-
+@Service
+@RequiredArgsConstructor
 public class FirebaseMessagingService {
-    @Bean
-    private FirebaseMessaging firebaseMessaging() throws IOException{
-        GoogleCredentials googleCredentials = GoogleCredentials
-                .fromStream(new ClassPathResource("firebase-service-account.json").getInputStream());
-        FirebaseOptions firebaseOptions = FirebaseOptions.builder()
-                .setCredentials(googleCredentials)
-                .build();
-        FirebaseApp app = FirebaseApp.initializeApp(firebaseOptions, "EZY");
-        return FirebaseMessaging.getInstance(app);
-    }
+
 }
